refactor: extract 5 route modules from server.py monolith
Move health, models, embeddings, MCP, and audio endpoints from
server.py (4025 lines) into dedicated route modules (484 lines total).

server.py: 4025 → 3564 lines (-11.5%)

Routes use FastAPI APIRouter and import server globals lazily to
avoid circular imports. Route modules are included at the bottom
of server.py after all globals are defined.

- routes/health.py: /health, /v1/status, /v1/cache/*
- routes/models.py: /v1/models
- routes/embeddings.py: /v1/embeddings
- routes/mcp_routes.py: /v1/mcp/*
- routes/audio.py: /v1/audio/*

All 433 unit tests pass. Lint clean.
