+++ vllm_mlx/routes/mcp_routes.py
@@ -1,0 +1,84 @@
+# SPDX-License-Identifier: Apache-2.0
+"""MCP (Model Context Protocol) endpoints."""
+
+from fastapi import APIRouter, Depends, HTTPException
+
+from ..api.models import (
+    MCPExecuteRequest,
+    MCPExecuteResponse,
+    MCPServerInfo,
+    MCPServersResponse,
+    MCPToolInfo,
+    MCPToolsResponse,
+)
+from ..server import verify_api_key
+
+router = APIRouter()
+
+
+@router.get("/v1/mcp/tools", dependencies=[Depends(verify_api_key)])
+async def list_mcp_tools() -> MCPToolsResponse:
+    """List all available MCP tools."""
+    from ..server import _mcp_manager
+
+    if _mcp_manager is None:
+        return MCPToolsResponse(tools=[], count=0)
+
+    tools = []
+    for tool in _mcp_manager.get_all_tools():
+        tools.append(
+            MCPToolInfo(
+                name=tool.full_name,
+                description=tool.description,
+                server=tool.server_name,
+                parameters=tool.input_schema,
+            )
+        )
+
+    return MCPToolsResponse(tools=tools, count=len(tools))
+
+
+@router.get("/v1/mcp/servers", dependencies=[Depends(verify_api_key)])
+async def list_mcp_servers() -> MCPServersResponse:
+    """Get status of all MCP servers."""
+    from ..server import _mcp_manager
+
+    if _mcp_manager is None:
+        return MCPServersResponse(servers=[])
+
+    servers = []
+    for status in _mcp_manager.get_server_status():
+        servers.append(
+            MCPServerInfo(
+                name=status.name,
+                state=status.state.value,
+                transport=status.transport.value,
+                tools_count=status.tools_count,
+                error=status.error,
+            )
+        )
+
+    return MCPServersResponse(servers=servers)
+
+
+@router.post("/v1/mcp/execute", dependencies=[Depends(verify_api_key)])
+async def execute_mcp_tool(request: MCPExecuteRequest) -> MCPExecuteResponse:
+    """Execute an MCP tool."""
+    from ..server import _mcp_manager
+
+    if _mcp_manager is None:
+        raise HTTPException(
+            status_code=503, detail="MCP not configured. Start server with --mcp-config"
+        )
+
+    result = await _mcp_manager.execute_tool(
+        request.tool_name,
+        request.arguments,
+    )
+
+    return MCPExecuteResponse(
+        tool_name=result.tool_name,
+        content=result.content,
+        is_error=result.is_error,
+        error_message=result.error_message,
+    )
